Make clippy happy

+++ src/ff/field.rs
@@ -43,10 +43,10 @@ pub trait Field: ArithmeticShare + From<u128> + Into<Self::Integer> {
     /// ## Errors
     /// Returns an error if buffer did not have enough capacity to store this field value
     fn serialize(&self, buf: &mut [u8]) -> io::Result<()> {
-        let raw_value = &self.as_u128().to_le_bytes()[..Self::SIZE_IN_BYTES as usize];
+        let raw_value = &self.as_u128().to_le_bytes()[..Self::SIZE_IN_BYTES];
 
         if buf.len() >= raw_value.len() {
-            buf[..Self::SIZE_IN_BYTES as usize].copy_from_slice(raw_value);
+            buf[..Self::SIZE_IN_BYTES].copy_from_slice(raw_value);
             Ok(())
         } else {
             let error_text = format!(
@@ -71,10 +71,9 @@ pub trait Field: ArithmeticShare + From<u128> + Into<Self::Integer> {
     /// ## Errors
     /// Returns an error if buffer did not have enough capacity left to read the field value.
     fn deserialize(buf_from: &mut [u8]) -> io::Result<Self> {
-        if Self::SIZE_IN_BYTES as usize <= buf_from.len() {
+        if Self::SIZE_IN_BYTES <= buf_from.len() {
             let mut buf_to = [0; 16]; // one day...
-            buf_to[..Self::SIZE_IN_BYTES as usize]
-                .copy_from_slice(&buf_from[..Self::SIZE_IN_BYTES as usize]);
+            buf_to[..Self::SIZE_IN_BYTES].copy_from_slice(&buf_from[..Self::SIZE_IN_BYTES]);
 
             Ok(Self::from(u128::from_le_bytes(buf_to)))
         } else {
